Install socketioxide and use it for Socket.IO server
