+++ AGENTS.md
@@ -1,0 +1,12 @@
+# AGENTS
+
+## Goal
+
+The project aims to port the Python voice changer server under `server/` to a Rust
+implementation in `server-rs/`. When adding features in Rust, mirror the Python
+code structure and behavior as closely as possible.
+
+## Conventions
+
+- Run `cargo fmt` and `cargo test --quiet` before committing.
+- Document new modules and describe differences from Python where applicable.
