+++ server-rs/Cargo.toml
@@ -17,6 +17,7 @@ tracing-subscriber = { version = "0.3", features = ["env-filter"] }
 futures-util = "0.3"
 tower-http = { version = "0.3", features = ["fs"] }
 reqwest = { version = "0.11", features = ["json"] }
+socketioxide = { version = "0.8", features = ["tracing", "state"] }
 
 [dev-dependencies]
 tower = "0.4"
