+++ server-rs/src/main.rs
@@ -18,6 +18,7 @@ use voice_changer_manager::VoiceChangerManager;
 mod mmvc_rest;
 mod mmvc_socketio_app;
 mod mmvc_socketio_server;
+mod socketio_stub;
 mod volume_extractor;
 use mmvc_rest::MMVCRest;
 use mmvc_socketio_app::MMVCSocketIOApp;
