+++ server-rs/src/socketio_stub.rs
@@ -1,0 +1,23 @@
+/// Placeholder module for future Socket.IO support.
+///
+/// The current Rust implementation only exposes a plain WebSocket endpoint.
+/// In the Python server, Socket.IO is used to send binary audio frames and
+/// receive processed results. Implementing the full protocol requires
+/// compatibility with Engine.IO and the Socket.IO framing format, which is
+/// non-trivial.
+///
+/// This file provides a minimal stub that documents the intended interface.
+
+pub struct SocketIOServer;
+
+impl SocketIOServer {
+    pub fn new() -> Self {
+        Self
+    }
+
+    /// Start the Socket.IO server. Currently unimplemented.
+    pub async fn start(&self) {
+        // TODO: implement Socket.IO handling using an appropriate crate
+        // or a custom implementation compatible with the Python version.
+    }
+}
